New function for EnumerateFilesAndFolders() over MediaDirectoryInfo
Adds a possibility to get all child items of directory at once.
It is useful when we need to calculate total number of items under
directory.
Unforatunatelly for that I had to make MediaFileSystemInfo normall class.
But maybe a beeter solution will be to rename this class to
MediaObjectSystemInfo and create new empty inherited class MediaObjectInfo?